Feat/30 : CreateMenu 기능 추가 (#34)
* feat : 메뉴 생성 구현

* feat : create menu 완료

* feat : @RequestBody에  Vaild 추가

* refactor : 응답 형식 변경

* refactor : 응답 형식 변경

* refactor : 변수 타입 변경

+++ src/main/java/com/example/toanyone/domain/menu/service/MenuService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 @Service
 public interface MenuService {
 
-    MenuDto.Response createMenu(Integer storeId, String name, String description,
+    MenuDto.Response createMenu(Long storeId, String name, String description,
                                 Integer price, MainCategory mainCategory, SubCategory subCategory);
 
 }
